+++ packages/api-core/src/features/security/apiKeys/GetApiKeyBySlug/GetApiKeyBySlugUseCase.ts
@@ -1,18 +1,10 @@
-import { Result } from "@webiny/feature/api";
 import { GetApiKeyBySlugUseCase as GetApiKeyBySlugUseCaseAbstraction } from "./abstractions.js";
 import { ApiKeysRepository } from "../shared/abstractions.js";
-import type { ApiKey } from "../shared/types.js";
 
 class GetApiKeyBySlugUseCaseImpl implements GetApiKeyBySlugUseCaseAbstraction.Interface {
-    private repository: ApiKeysRepository.Interface;
+    constructor(private repository: ApiKeysRepository.Interface) {}
 
-    constructor(repository: ApiKeysRepository.Interface) {
-        this.repository = repository;
-    }
-
-    async execute(
-        slug: string
-    ): Promise<Result<ApiKey | null, GetApiKeyBySlugUseCaseAbstraction.Error>> {
+    async execute(slug: string) {
         return this.repository.getBySlug(slug);
     }
 }
