+++ packages/api-core/src/features/security/apiKeys/GetApiKeyByToken/GetApiKeyByTokenUseCase.ts
@@ -1,18 +1,10 @@
-import { Result } from "@webiny/feature/api";
 import { GetApiKeyByTokenUseCase as GetApiKeyByTokenUseCaseAbstraction } from "./abstractions.js";
 import { ApiKeysRepository } from "../shared/abstractions.js";
-import type { ApiKey } from "../shared/types.js";
 
 class GetApiKeyByTokenUseCaseImpl implements GetApiKeyByTokenUseCaseAbstraction.Interface {
-    private repository: ApiKeysRepository.Interface;
+    constructor(private repository: ApiKeysRepository.Interface) {}
 
-    constructor(repository: ApiKeysRepository.Interface) {
-        this.repository = repository;
-    }
-
-    async execute(
-        token: string
-    ): Promise<Result<ApiKey | null, GetApiKeyByTokenUseCaseAbstraction.Error>> {
+    async execute(token: string) {
         return this.repository.getByToken(token);
     }
 }
