+++ packages/api-core/src/features/security/apiKeys/shared/types.ts
@@ -5,7 +5,7 @@ export interface ApiKey {
     name: string;
     slug: string;
     description: string;
-    token: string;
+    token: `wat_${string}`;
     permissions: SecurityPermission[];
     createdBy: CreatedBy;
     createdOn: string;
