+++ packages/sdk-frontend/webiny.config.js
