+++ packages/sdk-nextjs/src/webpack.ts
