+++ packages/sdk-nextjs/tsconfig.build.json
@@ -1,13 +1,13 @@
 {
-  "extends": "../../tsconfig.json",
-  "include": ["src", "__tests__"],
+  "extends": "../../tsconfig.build.json",
+  "include": ["src"],
   "references": [
-    { "path": "../cms-nextjs" },
-    { "path": "../content-sdk" },
-    { "path": "../website-builder-nextjs" }
+    { "path": "../cms-nextjs/tsconfig.build.json" },
+    { "path": "../sdk-frontend/tsconfig.build.json" },
+    { "path": "../website-builder-nextjs/tsconfig.build.json" }
   ],
   "compilerOptions": {
-    "rootDirs": ["./src", "./__tests__"],
+    "rootDir": "./src",
     "outDir": "./dist",
     "declarationDir": "./dist",
     "paths": {
@@ -15,8 +15,8 @@
       "~tests/*": ["./__tests__/*"],
       "@webiny/cms-nextjs/*": ["../cms-nextjs/src/*"],
       "@webiny/cms-nextjs": ["../cms-nextjs/src"],
-      "@webiny/content-sdk/*": ["../content-sdk/src/*"],
-      "@webiny/content-sdk": ["../content-sdk/src"],
+      "@webiny/sdk-frontend/*": ["../sdk-frontend/src/*"],
+      "@webiny/sdk-frontend": ["../sdk-frontend/src"],
       "@webiny/website-builder-nextjs/middleware": ["../website-builder-nextjs/src/middleware.js"],
       "@webiny/website-builder-nextjs/webpack": ["../website-builder-nextjs/src/webpack.js"],
       "@webiny/website-builder-nextjs/*": ["../website-builder-nextjs/src/*"],
