+++ packages/sdk-nextjs/tsconfig.json
@@ -1,13 +1,13 @@
 {
-  "extends": "../../tsconfig.build.json",
-  "include": ["src"],
+  "extends": "../../tsconfig.json",
+  "include": ["src", "__tests__"],
   "references": [
-    { "path": "../cms-nextjs/tsconfig.build.json" },
-    { "path": "../content-sdk/tsconfig.build.json" },
-    { "path": "../website-builder-nextjs/tsconfig.build.json" }
+    { "path": "../cms-nextjs" },
+    { "path": "../sdk-frontend" },
+    { "path": "../website-builder-nextjs" }
   ],
   "compilerOptions": {
-    "rootDir": "./src",
+    "rootDirs": ["./src", "./__tests__"],
     "outDir": "./dist",
     "declarationDir": "./dist",
     "paths": {
@@ -15,8 +15,8 @@
       "~tests/*": ["./__tests__/*"],
       "@webiny/cms-nextjs/*": ["../cms-nextjs/src/*"],
       "@webiny/cms-nextjs": ["../cms-nextjs/src"],
-      "@webiny/content-sdk/*": ["../content-sdk/src/*"],
-      "@webiny/content-sdk": ["../content-sdk/src"],
+      "@webiny/sdk-frontend/*": ["../sdk-frontend/src/*"],
+      "@webiny/sdk-frontend": ["../sdk-frontend/src"],
       "@webiny/website-builder-nextjs/middleware": ["../website-builder-nextjs/src/middleware.js"],
       "@webiny/website-builder-nextjs/webpack": ["../website-builder-nextjs/src/webpack.js"],
       "@webiny/website-builder-nextjs/*": ["../website-builder-nextjs/src/*"],
