+++ packages/webiny/src/api/ai-powerups.ts
@@ -1,2 +1,2 @@
 export { CmsGenerateEntryContentUseCase } from "@webiny/ai-powerups/api/features/CmsGenerateEntryContent/abstractions.js";
-export { GetSettingsUseCase } from "@webiny/ai-powerups/api/features/GetSettings/index.js";
+export { GetSettingsUseCase } from "@webiny/ai-powerups/api/features/GetSettings/abstractions.js";
