+++ packages/webiny/src/api/security/api-key.ts
@@ -10,6 +10,7 @@ export {
 } from "@webiny/api-core/features/security/apiKeys/DeleteApiKey/index.js";
 export { GetApiKeyUseCase } from "@webiny/api-core/features/security/apiKeys/GetApiKey/index.js";
 export { GetApiKeyByTokenUseCase } from "@webiny/api-core/features/security/apiKeys/GetApiKeyByToken/index.js";
+export { GetApiKeyBySlugUseCase } from "@webiny/api-core/features/security/apiKeys/GetApiKeyBySlug/index.js";
 export { ListApiKeysUseCase } from "@webiny/api-core/features/security/apiKeys/ListApiKeys/index.js";
 export {
     ApiKeyAfterUpdateEventHandler,
