+++ webiny.config.tsx
@@ -11,6 +11,8 @@ import { AiContentExtension } from "@/extensions/bulkActions/aiContent/AiContent
 export const Extensions = () => {
     return (
         <>
+            <Api.Extension src={"@/extensions/tenantTheme/TenantThemeExtension.ts"} />
+            <Api.Extension src={"/extensions/WebsiteBuilderApiKey.ts"} />
             {/* Admin 👇 */}
             <Admin.Extension src={"@/extensions/previewUrlModifier/index.tsx"} />
             {/*<Admin.Extension src={"@/extensions/fileUrlFormatter/index.tsx"} />*/}
